+++ lib/ansible/modules/network/f5/bigip_remote_syslog.py
@@ -24,6 +24,7 @@ options:
     description:
       - Specifies the IP address, or hostname, for the remote system to
         which the system sends log messages.
+    type: str
     required: True
   name:
     description:
@@ -31,6 +32,7 @@ options:
       - This option is required when multiple C(remote_host) with the same IP
         or hostname are present on the device.
       - If C(name) is not provided C(remote_host) is used by default.
+    type: str
     version_added: 2.8
   remote_port:
     description:
@@ -38,21 +40,24 @@ options:
         remote logging server.
       - When creating a remote syslog, if this parameter is not specified, the
         default value C(514) is used.
+    type: str
   local_ip:
     description:
       - Specifies the local IP address of the system that is logging. To
         provide no local IP, specify the value C(none).
       - When creating a remote syslog, if this parameter is not specified, the
         default value C(none) is used.
+    type: str
   state:
     description:
       - When C(present), guarantees that the remote syslog exists with the provided
         attributes.
       - When C(absent), removes the remote syslog from the system.
-    default: present
+    type: str
     choices:
       - absent
       - present
+    default: present
 extends_documentation_fragment: f5
 author:
   - Tim Rupp (@caphrim007)
@@ -100,11 +105,8 @@ try:
     from library.module_utils.network.f5.bigip import F5RestClient
     from library.module_utils.network.f5.common import F5ModuleError
     from library.module_utils.network.f5.common import AnsibleF5Parameters
-    from library.module_utils.network.f5.common import cleanup_tokens
     from library.module_utils.network.f5.common import f5_argument_spec
-    from library.module_utils.network.f5.common import exit_json
-    from library.module_utils.network.f5.common import fail_json
-    from library.module_utils.network.f5.common import compare_dictionary
+    from library.module_utils.network.f5.compare import compare_dictionary
     from library.module_utils.network.f5.common import is_valid_hostname
     from library.module_utils.network.f5.common import fq_name
     from library.module_utils.network.f5.ipaddress import is_valid_ip
@@ -112,11 +114,8 @@ except ImportError:
     from ansible.module_utils.network.f5.bigip import F5RestClient
     from ansible.module_utils.network.f5.common import F5ModuleError
     from ansible.module_utils.network.f5.common import AnsibleF5Parameters
-    from ansible.module_utils.network.f5.common import cleanup_tokens
     from ansible.module_utils.network.f5.common import f5_argument_spec
-    from ansible.module_utils.network.f5.common import exit_json
-    from ansible.module_utils.network.f5.common import fail_json
-    from ansible.module_utils.network.f5.common import compare_dictionary
+    from ansible.module_utils.network.f5.compare import compare_dictionary
     from ansible.module_utils.network.f5.common import is_valid_hostname
     from ansible.module_utils.network.f5.common import fq_name
     from ansible.module_utils.network.f5.ipaddress import is_valid_ip
@@ -258,7 +257,7 @@ class Difference(object):
 class ModuleManager(object):
     def __init__(self, *args, **kwargs):
         self.module = kwargs.pop('module', None)
-        self.client = kwargs.pop('client', None)
+        self.client = F5RestClient(**self.module.params)
         self.want = ModuleParameters(params=self.module.params)
         self.have = ApiParameters()
         self.changes = UsableChanges()
@@ -458,16 +457,12 @@ def main():
         supports_check_mode=spec.supports_check_mode
     )
 
-    client = F5RestClient(**module.params)
-
     try:
-        mm = ModuleManager(module=module, client=client)
+        mm = ModuleManager(module=module)
         results = mm.exec_module()
-        cleanup_tokens(client)
-        exit_json(module, results, client)
+        module.exit_json(**results)
     except F5ModuleError as ex:
-        cleanup_tokens(client)
-        fail_json(module, ex, client)
+        module.fail_json(msg=str(ex))
 
 
 if __name__ == '__main__':
