+++ test/units/modules/network/f5/test_bigip_gtm_monitor_external.py
@@ -113,9 +113,11 @@ class TestManager(unittest.TestCase):
             interval=20,
             timeout=30,
             partition='Common',
-            server='localhost',
-            password='password',
-            user='admin'
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
