+++ test/units/modules/network/f5/test_bigip_timer_policy.py
@@ -92,9 +92,11 @@ class TestManager(unittest.TestCase):
         set_module_args(dict(
             name='foo',
             description='my description',
-            server='localhost',
-            password='password',
-            user='admin'
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
