+++ test/units/modules/network/f5/test_bigip_vlan.py
@@ -107,15 +107,18 @@ class TestManager(unittest.TestCase):
         set_module_args(dict(
             name='somevlan',
             description='fakevlan',
-            server='localhost',
-            password='password',
-            user='admin',
-            partition='Common'
+            partition='Common',
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
             argument_spec=self.spec.argument_spec,
-            supports_check_mode=self.spec.supports_check_mode
+            supports_check_mode=self.spec.supports_check_mode,
+            mutually_exclusive=self.spec.mutually_exclusive
         )
 
         # Override methods to force specific logic in the module to happen
@@ -133,15 +136,18 @@ class TestManager(unittest.TestCase):
             name='somevlan',
             tagged_interface=['2.1'],
             tag=213,
-            server='localhost',
-            password='password',
-            user='admin',
-            partition='Common'
+            partition='Common',
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
             argument_spec=self.spec.argument_spec,
-            supports_check_mode=self.spec.supports_check_mode
+            supports_check_mode=self.spec.supports_check_mode,
+            mutually_exclusive=self.spec.mutually_exclusive
         )
 
         # Override methods to force specific logic in the module to happen
@@ -159,15 +165,18 @@ class TestManager(unittest.TestCase):
         set_module_args(dict(
             name='somevlan',
             untagged_interface=['2.1'],
-            server='localhost',
-            password='password',
-            user='admin',
-            partition='Common'
+            partition='Common',
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
             argument_spec=self.spec.argument_spec,
-            supports_check_mode=self.spec.supports_check_mode
+            supports_check_mode=self.spec.supports_check_mode,
+            mutually_exclusive=self.spec.mutually_exclusive
         )
 
         # Override methods to force specific logic in the module to happen
@@ -185,15 +194,18 @@ class TestManager(unittest.TestCase):
             name='somevlan',
             tagged_interface=['2.1', '1.1'],
             tag=213,
-            server='localhost',
-            password='password',
-            user='admin',
-            partition='Common'
+            partition='Common',
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
             argument_spec=self.spec.argument_spec,
-            supports_check_mode=self.spec.supports_check_mode
+            supports_check_mode=self.spec.supports_check_mode,
+            mutually_exclusive=self.spec.mutually_exclusive
         )
 
         # Override methods to force specific logic in the module to happen
@@ -211,15 +223,18 @@ class TestManager(unittest.TestCase):
         set_module_args(dict(
             name='somevlan',
             untagged_interface=['2.1', '1.1'],
-            server='localhost',
-            password='password',
-            user='admin',
             partition='Common',
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
             argument_spec=self.spec.argument_spec,
-            supports_check_mode=self.spec.supports_check_mode
+            supports_check_mode=self.spec.supports_check_mode,
+            mutually_exclusive=self.spec.mutually_exclusive
         )
 
         # Override methods to force specific logic in the module to happen
@@ -236,15 +251,18 @@ class TestManager(unittest.TestCase):
         set_module_args(dict(
             name='somevlan',
             untagged_interface=['2.1'],
-            server='localhost',
-            password='password',
-            user='admin',
             partition='Common',
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
             argument_spec=self.spec.argument_spec,
-            supports_check_mode=self.spec.supports_check_mode
+            supports_check_mode=self.spec.supports_check_mode,
+            mutually_exclusive=self.spec.mutually_exclusive
         )
 
         # Override methods to force specific logic in the module to happen
@@ -267,15 +285,18 @@ class TestManager(unittest.TestCase):
         set_module_args(dict(
             name='somevlan',
             tagged_interface=['2.1'],
-            server='localhost',
-            password='password',
-            user='admin',
             partition='Common',
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
             argument_spec=self.spec.argument_spec,
-            supports_check_mode=self.spec.supports_check_mode
+            supports_check_mode=self.spec.supports_check_mode,
+            mutually_exclusive=self.spec.mutually_exclusive
         )
 
         # Override methods to force specific logic in the module to happen
@@ -296,15 +317,18 @@ class TestManager(unittest.TestCase):
         set_module_args(dict(
             name='somevlan',
             description='changed_that',
-            server='localhost',
-            password='password',
-            user='admin',
             partition='Common',
+            provider=dict(
+                server='localhost',
+                password='password',
+                user='admin'
+            )
         ))
 
         module = AnsibleModule(
             argument_spec=self.spec.argument_spec,
-            supports_check_mode=self.spec.supports_check_mode
+            supports_check_mode=self.spec.supports_check_mode,
+            mutually_exclusive=self.spec.mutually_exclusive
         )
 
         # Override methods to force specific logic in the module to happen
